Add test of App component
